refactor and split out code for profile page. update convex functions to allow all users to have publicly available pages
