+++ apps/web/app/(app)/talent/profile/preview/page.tsx
@@ -1,16 +1,19 @@
-import { ProfileCard } from './profile-card'
+import { ProfileCard } from '@/components/features/profile/profile-card'
 import { me } from '@/lib/server/users'
 
-import { PreviewTabs } from './tab-section/preview-tabs'
-import { ScrollCollapse } from './scroll-collapse'
+import { ScrollCollapse } from '@/components/features/profile/scroll-collapse'
+import { TabSection } from '@/components/features/profile/tab-section'
+import { getMyResume } from '@/lib/server/resumes'
+
 export default async function ProfilePage() {
   const user = await me()
+  const resume = await getMyResume()
   return (
     <div className="flex flex-col gap-2 ">
       <ScrollCollapse collapsedTitle={`${user.firstName} ${user.lastName}`}>
         <ProfileCard user={user} />
       </ScrollCollapse>
-      <PreviewTabs />
+      <TabSection user={user} resume={resume} />
     </div>
   )
 }
