--- apps/web/app/(app)/talent/profile/preview/tab-section/preview-tabs.css
@@ -1,7 +1,0 @@
-.hover\:underline:hover{
-  text-decoration: none !important;
-}
-
-.\[\&\[data-state\=open\]\>svg\]\:rotate-45[data-state=open]>svg{
-  transform: rotate(90deg);
-}
