+++ apps/web/lib/utils.ts
@@ -1,5 +1,6 @@
 import { type ClassValue, clsx } from 'clsx'
 import { extendTailwindMerge } from 'tailwind-merge'
+import { toDate, differenceInYears, endOfToday } from 'date-fns'
 
 import { customFontSizes } from '../tailwind.config'
 
@@ -30,4 +31,25 @@ export const UNITS = {
 export function formatSizeValue(value: number, unit?: keyof typeof UNITS) {
   if (!value) return value
   return `${value}${unit ? UNITS[unit] : ''}`
+}
+
+export function locationToString(location?: {
+  city: string
+  state: string
+  country: string
+  address?: string
+  zipCode?: string
+}) {
+  if (!location) {
+    return ''
+  }
+  return `${location.address} ${location.city}, ${location.state} ${location.zipCode}`.trim()
+}
+
+export function calculateAge(dateOfBirth?: string | null) {
+  if (!dateOfBirth) {
+    return
+  }
+  const dob = toDate(dateOfBirth)
+  return differenceInYears(endOfToday(), dob)
 }
