+++ packages/backend/convex/agencies.ts
@@ -3,6 +3,9 @@ import { authMutation, authQuery } from './util'
 
 import { crud } from 'convex-helpers/server'
 import { Agencies } from './schema'
+import { Doc } from './_generated/dataModel'
+
+export type AgencyDoc = Doc<'agencies'>
 
 export const { read } = crud(Agencies, query, mutation)
